+++ src/tools/string.hpp
@@ -12,7 +12,7 @@ std::string nextToken(const std::string& string, char separator,
         size_t& position);
 
 std::string intToString(int i, unsigned radix = 10);
-std::string floatToString(float i, int decimals);
+std::string floatToString(double i, int decimals);
 
 class Join {
 public:
